+++ app/components/albumcard.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script setup lang="ts">
-import type { SpotifyAlbumItem } from '~/lib/types/spotify/album/spotify-album.types';
+import type { SpotifyAlbumItem } from '~/lib/types/spotify/album/spotify-album-types';
 
 
 const { name, images, artists } = defineProps<Partial<SpotifyAlbumItem>>();
